refactor(tabs): extract to another components
